+++ HISTORY.md
@@ -1,5 +1,15 @@
 # history
 
+## 0.2.2
+
+* `CHANGED` Do not parse uxcore modules in babel-loader
+* `CHANGED` sourceMap use SourceMapDevToolPlugin
+* `NEW` add happyPack
+* `NEW` open default browser when server start
+* `NEW` add process bar of webpack building
+* `NEW` add portInUse detect
+* `NEW` show true ip in console
+
 ## 0.2.1
 
 * `FIX` fix publish bug